Prevent out of bound reading in smartstrcat and ParseSmiles_smartstrcat.
These overruns occurred when ReparseUrl finished at the end
of the source string, in this case source pointer pointed to '\0',
but it was still unconditionally incremented and the next check
in ``while (*s == 0)`` read after the end of the source buffer.

This bug was detected by both ASan and Valgrind.