Finally fix an off-by-one bug that caused profindex.idx corruption.
Each time a user was deleted, the MSB of the last entry in the modified
hash block was mistakenly zeroed. E.g. 0x107A3F4 becomes 0x07A3F4,
which is invalid and may even overlap another user profile. This bug was
present since the very beginning of this codebase, so it may be 20+ y.o.

Now I have to implement a fix for all affected entries...